Combined and deleted some files to make things easier to edit and follow
