--- cap1/src/components/Profile.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const Profile = () => {
-  return (
-    <h3>Profile</h3>
-  );
-};
-
-export default Profile;
